+++ c/atan2.c
@@ -12,14 +12,14 @@ namespace codee {
 extern "C" {
 #endif
 
-int atan2_s (float *Y, const float *X1, const float *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor);
-int atan2_d (double *Y, const double *X1, const double *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor);
+int atan2_s (float *Y, const float *X1, const float *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor);
+int atan2_d (double *Y, const double *X1, const double *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor);
 
-int atan2_inplace_s (float *X1, const float *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor);
-int atan2_inplace_d (double *X1, const double *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor);
+int atan2_inplace_s (float *X1, const float *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor);
+int atan2_inplace_d (double *X1, const double *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor);
 
 
-int atan2_s (float *Y, const float *X1, const float *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor)
+int atan2_s (float *Y, const float *X1, const float *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor)
 {
     if (R1<0) { fprintf(stderr,"error in atan2_s: R1 (num rows X1) must be nonnegative\n"); return 1; }
     if (C1<0) { fprintf(stderr,"error in atan2_s: C1 (num cols X1) must be nonnegative\n"); return 1; }
@@ -55,15 +55,15 @@ int atan2_s (float *Y, const float *X1, const float *X2, const int R1, const int
         const int RCS1 = R1*C1*(S1>1), RCS2 = R2*C2*(S2>1);
         const int RC1 = R1*(C1>1), RC2 = R2*(C2>1);
         const int r1i = (R1>1), r2i = (R2>1);
-        for (int h=0; h<H; h++)
+        for (size_t h=0; h<H; h++)
         {
-            for (int s=0; s<S; s++)
+            for (size_t s=0; s<S; s++)
             {
-                for (int c=0; c<C; c++)
+                for (size_t c=0; c<C; c++)
                 {
                     n1 = h*RCSH1 + s*RCS1 + c*RC1;
                     n2 = h*RCSH2 + s*RCS2 + c*RC2;
-                    for (int r=0; r<R; r++)
+                    for (size_t r=0; r<R; r++)
                     {
                         Y[n] = atan2f(X1[n1],X2[n2]);
                         n++; n1 += r1i; n2 += r2i;
@@ -78,15 +78,15 @@ int atan2_s (float *Y, const float *X1, const float *X2, const int R1, const int
         const int HSC1 = H1*S1*(C1>1), HSC2 = H2*S2*(C2>1);
         const int HS1 = H1*(S1>1), HS2 = H2*(S2>1);
         const int h1i = (H1>1), h2i = (H2>1);
-        for (int r=0; r<R; r++)
+        for (size_t r=0; r<R; r++)
         {
-            for (int c=0; c<C; c++)
+            for (size_t c=0; c<C; c++)
             {
-                for (int s=0; s<S; s++)
+                for (size_t s=0; s<S; s++)
                 {
                     n1 = r*HSCR1 + c*HSC1 + s*HS1;
                     n2 = r*HSCR2 + c*HSC2 + s*HS2;
-                    for (int h=0; h<H; h++)
+                    for (size_t h=0; h<H; h++)
                     {
                         Y[n] = atan2f(X1[n1],X2[n2]);
                         n++; n1 += h1i; n2 += h2i;
@@ -100,7 +100,7 @@ int atan2_s (float *Y, const float *X1, const float *X2, const int R1, const int
 }
 
 
-int atan2_d (double *Y, const double *X1, const double *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor)
+int atan2_d (double *Y, const double *X1, const double *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor)
 {
     if (R1<0) { fprintf(stderr,"error in atan2_d: R1 (num rows X1) must be nonnegative\n"); return 1; }
     if (C1<0) { fprintf(stderr,"error in atan2_d: C1 (num cols X1) must be nonnegative\n"); return 1; }
@@ -136,15 +136,15 @@ int atan2_d (double *Y, const double *X1, const double *X2, const int R1, const 
         const int RCS1 = R1*C1*(S1>1), RCS2 = R2*C2*(S2>1);
         const int RC1 = R1*(C1>1), RC2 = R2*(C2>1);
         const int r1i = (R1>1), r2i = (R2>1);
-        for (int h=0; h<H; h++)
+        for (size_t h=0; h<H; h++)
         {
-            for (int s=0; s<S; s++)
+            for (size_t s=0; s<S; s++)
             {
-                for (int c=0; c<C; c++)
+                for (size_t c=0; c<C; c++)
                 {
                     n1 = h*RCSH1 + s*RCS1 + c*RC1;
                     n2 = h*RCSH2 + s*RCS2 + c*RC2;
-                    for (int r=0; r<R; r++)
+                    for (size_t r=0; r<R; r++)
                     {
                         Y[n] = atan2(X1[n1],X2[n2]);
                         n++; n1 += r1i; n2 += r2i;
@@ -159,15 +159,15 @@ int atan2_d (double *Y, const double *X1, const double *X2, const int R1, const 
         const int HSC1 = H1*S1*(C1>1), HSC2 = H2*S2*(C2>1);
         const int HS1 = H1*(S1>1), HS2 = H2*(S2>1);
         const int h1i = (H1>1), h2i = (H2>1);
-        for (int r=0; r<R; r++)
+        for (size_t r=0; r<R; r++)
         {
-            for (int c=0; c<C; c++)
+            for (size_t c=0; c<C; c++)
             {
-                for (int s=0; s<S; s++)
+                for (size_t s=0; s<S; s++)
                 {
                     n1 = r*HSCR1 + c*HSC1 + s*HS1;
                     n2 = r*HSCR2 + c*HSC2 + s*HS2;
-                    for (int h=0; h<H; h++)
+                    for (size_t h=0; h<H; h++)
                     {
                         Y[n] = atan2(X1[n1],X2[n2]);
                         n++; n1 += h1i; n2 += h2i;
@@ -181,7 +181,7 @@ int atan2_d (double *Y, const double *X1, const double *X2, const int R1, const 
 }
 
 
-int atan2_inplace_s (float *X1, const float *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor)
+int atan2_inplace_s (float *X1, const float *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor)
 {
     if (R1<0) { fprintf(stderr,"error in atan2_inplace_s: R1 (num rows X1) must be nonnegative\n"); return 1; }
     if (C1<0) { fprintf(stderr,"error in atan2_inplace_s: C1 (num cols X1) must be nonnegative\n"); return 1; }
@@ -213,14 +213,14 @@ int atan2_inplace_s (float *X1, const float *X2, const int R1, const int C1, con
     else if (iscolmajor)
     {
         const int RCSH2 = R2*C2*S2*(H2>1), RCS2 = R2*C2*(S2>1), RC2 = R2*(C2>1), r2i = (R2>1);
-        for (int h=0; h<H; h++)
+        for (size_t h=0; h<H; h++)
         {
-            for (int s=0; s<S; s++)
+            for (size_t s=0; s<S; s++)
             {
-                for (int c=0; c<C; c++)
+                for (size_t c=0; c<C; c++)
                 {
                     n2 = h*RCSH2 + s*RCS2 + c*RC2;
-                    for (int r=0; r<R; r++)
+                    for (size_t r=0; r<R; r++)
                     {
                         X1[n] = atan2f(X1[n],X2[n2]);
                         n++; n2 += r2i;
@@ -232,14 +232,14 @@ int atan2_inplace_s (float *X1, const float *X2, const int R1, const int C1, con
     else
     {
         const int HSCR2 = H2*S2*C2*(R2>1), HSC2 = H2*S2*(C2>1), HS2 = H2*(S2>1), h2i = (H2>1);
-        for (int r=0; r<R; r++)
+        for (size_t r=0; r<R; r++)
         {
-            for (int c=0; c<C; c++)
+            for (size_t c=0; c<C; c++)
             {
-                for (int s=0; s<S; s++)
+                for (size_t s=0; s<S; s++)
                 {
                     n2 = r*HSCR2 + c*HSC2 + s*HS2;
-                    for (int h=0; h<H; h++)
+                    for (size_t h=0; h<H; h++)
                     {
                         X1[n] = atan2f(X1[n],X2[n2]);
                         n++; n2 += h2i;
@@ -255,7 +255,7 @@ int atan2_inplace_s (float *X1, const float *X2, const int R1, const int C1, con
 }
 
 
-int atan2_inplace_d (double *X1, const double *X2, const int R1, const int C1, const int S1, const int H1, const int R2, const int C2, const int S2, const int H2, const char iscolmajor)
+int atan2_inplace_d (double *X1, const double *X2, const size_t R1, const size_t C1, const size_t S1, const size_t H1, const size_t R2, const size_t C2, const size_t S2, const size_t H2, const char iscolmajor)
 {
     if (R1<0) { fprintf(stderr,"error in atan2_inplace_d: R1 (num rows X1) must be nonnegative\n"); return 1; }
     if (C1<0) { fprintf(stderr,"error in atan2_inplace_d: C1 (num cols X1) must be nonnegative\n"); return 1; }
@@ -285,14 +285,14 @@ int atan2_inplace_d (double *X1, const double *X2, const int R1, const int C1, c
     else if (iscolmajor)
     {
         const int RCSH2 = R2*C2*S2*(H2>1), RCS2 = R2*C2*(S2>1), RC2 = R2*(C2>1), r2i = (R2>1);
-        for (int h=0; h<H; h++)
+        for (size_t h=0; h<H; h++)
         {
-            for (int s=0; s<S; s++)
+            for (size_t s=0; s<S; s++)
             {
-                for (int c=0; c<C; c++)
+                for (size_t c=0; c<C; c++)
                 {
                     n2 = h*RCSH2 + s*RCS2 + c*RC2;
-                    for (int r=0; r<R; r++)
+                    for (size_t r=0; r<R; r++)
                     {
                         X1[n] = atan2(X1[n],X2[n2]);
                         n++; n2 += r2i;
@@ -304,14 +304,14 @@ int atan2_inplace_d (double *X1, const double *X2, const int R1, const int C1, c
     else
     {
         const int HSCR2 = H2*S2*C2*(R2>1), HSC2 = H2*S2*(C2>1), HS2 = H2*(S2>1), h2i = (H2>1);
-        for (int r=0; r<R; r++)
+        for (size_t r=0; r<R; r++)
         {
-            for (int c=0; c<C; c++)
+            for (size_t c=0; c<C; c++)
             {
-                for (int s=0; s<S; s++)
+                for (size_t s=0; s<S; s++)
                 {
                     n2 = r*HSCR2 + c*HSC2 + s*HS2;
-                    for (int h=0; h<H; h++)
+                    for (size_t h=0; h<H; h++)
                     {
                         X1[n] = atan2(X1[n],X2[n2]);
                         n++; n2 += h2i;
