+++ c/cnt.c
@@ -9,18 +9,18 @@ namespace codee {
 extern "C" {
 #endif
 
-int cnt_s (float *Y, const float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
-int cnt_d (double *Y, const double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
-int cnt_c (float *Y, const float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
-int cnt_z (double *Y, const double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
+int cnt_s (float *Y, const float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
+int cnt_d (double *Y, const double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
+int cnt_c (float *Y, const float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
+int cnt_z (double *Y, const double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
 
-int cnt_inplace_s (float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
-int cnt_inplace_d (double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
-int cnt_inplace_c (float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
-int cnt_inplace_z (double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor);
+int cnt_inplace_s (float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
+int cnt_inplace_d (double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
+int cnt_inplace_c (float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
+int cnt_inplace_z (double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor);
 
 
-int cnt_s (float *Y, const float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_s (float *Y, const float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_s: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_s: C (ncols X) must be positive\n"); return 1; }
@@ -50,7 +50,7 @@ int cnt_s (float *Y, const float *X, const int R, const int C,const int S, const
 }
 
 
-int cnt_d (double *Y, const double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_d (double *Y, const double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_d: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_d: C (ncols X) must be positive\n"); return 1; }
@@ -77,7 +77,7 @@ int cnt_d (double *Y, const double *X, const int R, const int C,const int S, con
 }
 
 
-int cnt_c (float *Y, const float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_c (float *Y, const float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_c: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_c: C (ncols X) must be positive\n"); return 1; }
@@ -104,7 +104,7 @@ int cnt_c (float *Y, const float *X, const int R, const int C,const int S, const
 }
 
 
-int cnt_z (double *Y, const double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_z (double *Y, const double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_z: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_z: C (ncols X) must be positive\n"); return 1; }
@@ -131,7 +131,7 @@ int cnt_z (double *Y, const double *X, const int R, const int C,const int S, con
 }
 
 
-int cnt_inplace_s (float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_inplace_s (float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_inplace_s: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_inplace_s: C (ncols X) must be positive\n"); return 1; }
@@ -163,7 +163,7 @@ int cnt_inplace_s (float *X, const int R, const int C,const int S, const int H, 
 }
 
 
-int cnt_inplace_d (double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_inplace_d (double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_inplace_d: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_inplace_d: C (ncols X) must be positive\n"); return 1; }
@@ -192,7 +192,7 @@ int cnt_inplace_d (double *X, const int R, const int C,const int S, const int H,
 }
 
 
-int cnt_inplace_c (float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_inplace_c (float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_inplace_c: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_inplace_c: C (ncols X) must be positive\n"); return 1; }
@@ -221,7 +221,7 @@ int cnt_inplace_c (float *X, const int R, const int C,const int S, const int H, 
 }
 
 
-int cnt_inplace_z (double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor)
+int cnt_inplace_z (double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor)
 {
     if (R<1) { fprintf(stderr,"error in cnt_inplace_z: R (nrows X) must be positive\n"); return 1; }
     if (C<1) { fprintf(stderr,"error in cnt_inplace_z: C (ncols X) must be positive\n"); return 1; }
