+++ c/coeff_var.c
@@ -11,11 +11,11 @@ namespace codee {
 extern "C" {
 #endif
 
-int coeff_var_s (float *Y, float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor, const char biased);
-int coeff_var_d (double *Y, double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor, const char biased);
+int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor, const char biased);
+int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor, const char biased);
 
 
-int coeff_var_s (float *Y, float *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor, const char biased)
+int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor, const char biased)
 {
     const float o = 1.0f;
     float m, den;
@@ -31,18 +31,18 @@ int coeff_var_s (float *Y, float *X, const int R, const int C,const int S, const
         {
             for (c=0; c<C; c++)
             {
-                m = cblas_sdot(R,&X[c*R],1,&o,0) / R;
-                cblas_saxpy(R,-m,&o,0,&X[c*R],1);
-                Y[c] = sqrtf(cblas_snrm2(R,&X[c*R],1)/den) / m;
+                m = cblas_sdot((int)R,&X[c*R],1,&o,0) / R;
+                cblas_saxpy((int)R,-m,&o,0,&X[c*R],1);
+                Y[c] = sqrtf(cblas_snrm2((int)R,&X[c*R],1)/den) / m;
             }
         }
         else
         {
             for (c=0; c<C; c++)
             {
-                m = cblas_sdot(R,&X[c],C,&o,0) / R;
-                cblas_saxpy(R,-m,&o,0,&X[c],C);
-                Y[c] = sqrtf(cblas_snrm2(R,&X[c],C)/den) / m;
+                m = cblas_sdot((int)R,&X[c],(int)C,&o,0) / R;
+                cblas_saxpy((int)R,-m,&o,0,&X[c],(int)C);
+                Y[c] = sqrtf(cblas_snrm2((int)R,&X[c],(int)C)/den) / m;
             }
         }
     }
@@ -53,18 +53,18 @@ int coeff_var_s (float *Y, float *X, const int R, const int C,const int S, const
         {
             for (r=0; r<R; r++)
             {
-                m = cblas_sdot(C,&X[r],R,&o,0) / C;
-                cblas_saxpy(C,-m,&o,0,&X[r],R);
-                Y[r] = sqrtf(cblas_snrm2(C,&X[r],R)/den) / m;
+                m = cblas_sdot((int)C,&X[r],(int)R,&o,0) / C;
+                cblas_saxpy((int)C,-m,&o,0,&X[r],(int)R);
+                Y[r] = sqrtf(cblas_snrm2((int)C,&X[r],(int)R)/den) / m;
             }
         }
         else
         {
             for (r=0; r<R; r++)
             {
-                m = cblas_sdot(C,&X[r*C],1,&o,0) / C;
-                cblas_saxpy(C,-m,&o,0,&X[r*C],1);
-                Y[r] = sqrtf(cblas_snrm2(C,&X[r*C],1)/den) / m;
+                m = cblas_sdot((int)C,&X[r*C],1,&o,0) / C;
+                cblas_saxpy((int)C,-m,&o,0,&X[r*C],1);
+                Y[r] = sqrtf(cblas_snrm2((int)C,&X[r*C],1)/den) / m;
             }
         }
     }
@@ -77,7 +77,7 @@ int coeff_var_s (float *Y, float *X, const int R, const int C,const int S, const
 }
 
 
-int coeff_var_d (double *Y, double *X, const int R, const int C,const int S, const int H, const int dim, const char iscolmajor, const char biased)
+int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size_t S, const size_t H, const int dim, const char iscolmajor, const char biased)
 {
     const double o = 1.0;
     double m, den;
@@ -93,18 +93,18 @@ int coeff_var_d (double *Y, double *X, const int R, const int C,const int S, con
         {
             for (c=0; c<C; c++)
             {
-                m = cblas_ddot(R,&X[c*R],1,&o,0) / R;
-                cblas_daxpy(R,-m,&o,0,&X[c*R],1);
-                Y[c] = sqrt(cblas_dnrm2(R,&X[c*R],1)/den) / m;
+                m = cblas_ddot((int)R,&X[c*R],1,&o,0) / R;
+                cblas_daxpy((int)R,-m,&o,0,&X[c*R],1);
+                Y[c] = sqrt(cblas_dnrm2((int)R,&X[c*R],1)/den) / m;
             }
         }
         else
         {
             for (c=0; c<C; c++)
             {
-                m = cblas_ddot(R,&X[c],C,&o,0) / R;
-                cblas_daxpy(R,-m,&o,0,&X[c],C);
-                Y[c] = sqrt(cblas_dnrm2(R,&X[c],C)/den) / m;
+                m = cblas_ddot((int)R,&X[c],(int)C,&o,0) / R;
+                cblas_daxpy((int)R,-m,&o,0,&X[c],(int)C);
+                Y[c] = sqrt(cblas_dnrm2((int)R,&X[c],(int)C)/den) / m;
             }
         }
     }
@@ -115,18 +115,18 @@ int coeff_var_d (double *Y, double *X, const int R, const int C,const int S, con
         {
             for (r=0; r<R; r++)
             {
-                m = cblas_ddot(C,&X[r],R,&o,0) / C;
-                cblas_daxpy(C,-m,&o,0,&X[r],R);
-                Y[r] = sqrt(cblas_dnrm2(C,&X[r],R)/den) / m;
+                m = cblas_ddot((int)C,&X[r],(int)R,&o,0) / C;
+                cblas_daxpy((int)C,-m,&o,0,&X[r],(int)R);
+                Y[r] = sqrt(cblas_dnrm2((int)C,&X[r],(int)R)/den) / m;
             }
         }
         else
         {
             for (r=0; r<R; r++)
             {
-                m = cblas_ddot(C,&X[r*C],1,&o,0) / C;
-                cblas_daxpy(C,-m,&o,0,&X[r*C],1);
-                Y[r] = sqrt(cblas_dnrm2(C,&X[r*C],1)/den) / m;
+                m = cblas_ddot((int)C,&X[r*C],1,&o,0) / C;
+                cblas_daxpy((int)C,-m,&o,0,&X[r*C],1);
+                Y[r] = sqrt(cblas_dnrm2((int)C,&X[r*C],1)/den) / m;
             }
         }
     }
